改造测试

+++ myimage/myimage/图片/Models/ArticleModel.h
@@ -13,13 +13,13 @@ NS_ASSUME_NONNULL_BEGIN
 @interface ArticleModel : NSObject
 
 @property (nonatomic, assign)int article_id;
-@property (nonatomic, assign)int website_id;
+@property (nonatomic, assign)NSInteger website_id;
 @property (nonatomic, assign)int category_id;
 @property (nonatomic, copy)NSString *name;
 @property (nonatomic, copy)NSString *detail_url;
 @property (nonatomic, copy)NSString *img_url;
 @property (nonatomic, assign)int has_done;
-@property(nonatomic, assign) int aid;
+@property(nonatomic, assign) NSInteger aid;
 @property (nonatomic, assign)int is_delete;
 
 -(instancetype)initWithArticleCollectModel:(ArticleCollectModel *)model;
